Tests/LibWeb/Text/expected/indexeddb-abort-in-success.txt

tx.abort() called
put.error: AbortError
second put failed: TransactionInactiveError: Transaction is not active while running add/put
tx.abort
DONE
